requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
free-proxy==1.0.2
idna==2.10
lxml==4.6.3
multidict==5.1.0
python-dotenv==0.15.0
pytz==2021.1
random-user-agent==1.0.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
websockets==8.1
yarl==1.6.3